fix caption trails on text change
gifs are optimized and only part of the image is re-written at each new
frame. The optimization is performed by ffmpeg and we are adding text
onto this optimized image therefore we are breaking the
optimization. This is fixed by recreating a new gif with ffmpeg each
time the text changes.